bt/requirements.txt
yfinance
pandas
toml
tqdm
requests
backtrader
llama-cpp-python
